fixes for #1155 PR review
